Adaptando el codigo para que funcione en la clase Simulador
